requirements.txt
pip
transformers
accelerate
datasets
sentencepiece
PyPDF2
tqdm
wandb
torch
python-dotenv
